% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AvgMargin.R
\name{AvgMargin}
\alias{AvgMargin}
\title{Compute average marginal effects}
\usage{
AvgMargin(data, model, voi, range = c(0:10))
}
\arguments{
\item{data}{A dateframe with voi}

\item{model}{lm or glm model}

\item{voi}{Variable of interest to be calculated.}

\item{range}{The value of voi to be predicted. For factor, use levels(data$voi) or specify factor levels}
}
\value{
dataframe of average predicted value at sepecified levels
}
\description{
\code{AvgMargin} computes average predicted values based on the given data at specified level of voi
}
\keyword{margin}
