% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BamCountConfig.R
\docType{class}
\name{BamCountConfig-class}
\alias{BamCountConfig}
\alias{BamCountConfig-class}
\alias{countConfigPairedEnd}
\alias{countConfigPairedEnd,ANY-method}
\alias{countConfigSingleEnd}
\alias{countConfigSingleEnd,ANY-method}
\alias{getFilter}
\alias{getFilter,BamCountConfig-method}
\title{Container for counting with bamsignals}
\usage{
\S4method{countConfigPairedEnd}{ANY}(binsize = 250L, mapq = 20L,
  filteredFlag = 1024L, shift = 0L, midpoint = T, tlenfilter = c(70L,
  200L))

\S4method{countConfigSingleEnd}{ANY}(binsize = 250L, mapq = 20L,
  filteredFlag = 1024L, shift = 0L)

\S4method{getFilter}{BamCountConfig}(obj)
}
\value{
return values are described in the Methods section.
}
\description{
This S4 class is a small wrapper for a configuration on obtaining counts
from bamfiles with bamsignals::bamProfile().
}
\section{Methods (by generic)}{
\itemize{
\item \code{countConfigPairedEnd}: Setup a paired end counting configuration

\item \code{countConfigSingleEnd}: Setup a single end counting configuration

\item \code{getFilter}: Get the bamsignals relevant filter
}}
\section{Fields}{

\describe{
\item{\code{type}}{A \code{character()} representing the type of bamfile
("paired.end", "single.end").}

\item{\code{chromosomes}}{GenomicRanges object used to specify the chromosomes.}

\item{\code{binsize}}{An \code{integer()} giving the binsize to count in (in bp).}

\item{\code{mapq}}{discard reads with mapping quality strictly lower than this
parameter. The value 0 ensures that no read will be discarded, the value 254
that only reads with the highest possible mapping quality will be considered.}

\item{\code{filteredFlag}}{An \code{integer()} to filter for when counting reads (SAMFLAG).
For example, 1024 filters out marked duplicates (default), see
\url{https://broadinstitute.github.io/picard/explain-flags.html}}

\item{\code{shift}}{shift the read position by a user-defined number of basepairs.
This can be handy in the analysis of chip-seq data.}

\item{\code{midpoint}}{A \code{logical()} indicating whether fragment midpoints
instead of 5'-ends should be counted in paired end data.}

\item{\code{tlenfilter}}{A filter on fragment length as estimated from alignment
in paired end experiments (TLEN). If set to \code{c(min,max)} only reads are
considered where \code{min <= TLEN <= max}. If \code{paired.end=="ignore"},
this argument is set to \code{NULL} and no filtering is done. If
\code{paired.end!="ignore"}, this argument defaults to \code{c(0,1000)}.}
}}
\seealso{
\code{\link{normr-methods}} for the functions that require this
object.\cr
See \code{\link{bamsignals}}-package for counting in bam files.
}

