% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wrappers.R
\name{bin.genome}
\alias{bin.genome}
\title{Create bins for a given genome annotation.}
\usage{
bin.genome(genome, bin.size = 300)
}
\arguments{
\item{genome}{A \link{data.frame} consisting of two columns. First column
gives sequence names. Second column gives chromosome lengths. This
information can be retrieved via the UCSC chromSizes table.}

\item{bin.size}{The size of the bins in bp.}
}
\value{
\link{GenomicRanges}-object specifying bins for genome
}
\description{
Computes non-overlapping bins for given chromosome names and lengths.
}

