% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{computeEnrichment}
\alias{computeEnrichment}
\title{Get normalized enrichment from a diffR fit}
\usage{
computeEnrichment(r, s, posteriors, B = 0L, nthreads = 1L)
}
\arguments{
\item{r}{vector of counts in control/condition1}

\item{s}{vector of counts in treatment/condition2}

\item{posteriors}{posterior matrix as computed by a normR routine}

\item{B}{column index of background component in posteriors (DEFAULT=0)}
}
\value{
a numeric with enrichment values in log space
}
\description{
Get normalized enrichment from a diffR fit
}

