% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityQuartile.R
\name{densityQuartile}
\alias{densityQuartile}
\title{Value of VOI at the Quartile of Its Density Distribution}
\usage{
densityQuartile(data, dv, bandwidth = 1, geotype = "Urban")
}
\arguments{
\item{data}{A dataset with var "psutype" and "psuname".}

\item{dv}{Variable of interest to be calculated.}

\item{bandwidth}{see \link[stat]{density} bw.}

\item{geotype}{Use factor lable of "psutype" variable.}
}
\value{
A martix contrains value of VOI at the qualtile of density.
}
\description{
\code{densityQuartile} calculate value of VOI at the quartile of density.
It is written for supplying value to be specified in the boxplot.
The boxplot can be combined with density plot.
The calculation following 3 steps:
  Step1 find the density distribution.
  Step2 find the value of the data point at 0 25 50 75 100 percentile of the distribution.
  Step3 take the value back to the density plot.
}
\keyword{density}
