% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wrappers.R
\name{diffcall}
\alias{diffcall}
\title{Call differences in two NGS experiments.}
\usage{
diffcall(treatment.1, treatment.2, genome, bin.size = 300, models = 4,
  eps = 0.001, p.values = T, procs = 1, mapqual = 20, shift = 0,
  paired.end = "ignore", verbose = T)
}
\arguments{
\item{treatment.1}{A \link{numeric} vector of treatment.1 counts or a
\link{character} pointing to the treatment.1 bam file. In the latter case an
\code{treatment.1}.bai index file should exist in the same folder. Should be
consistent with treatment.2.}

\item{treatment.2}{A \link{numeric} vector of treatment.2 counts or a
\link{character} pointing to the treatment.2 bam file. In the latter case an
\code{treatment.2}.bai index file should exist in the same folder. Should be
consistent with treatment.1.}

\item{genome}{A \link{data.frame} consisting of two columns. First column
gives sequence names. Second column gives chromosome lengths. This
information can be retrieved via the UCSC chromSizes table.}

\item{bin.size}{Width of genomic bins in bp.}

\item{models}{Number of model components. Note that this should assemble to
1 background components and at least 2 treatment components. Default: 4}

\item{eps}{Threshold for EM convergence.}

\item{p.values}{Flag for P value computation.}

\item{procs}{Number of threads to use in parallel::mclapply()}

\item{mapqual}{discard reads with mapping quality strictly lower than this
parameter. The value 0 ensures that no read will be discarded, the value 254
that only reads with the highest possible mapping quality will be considered.}

\item{shift}{shift the read position by a user-defined number of basepairs.
This can be handy in the analysis of chip-seq data.}

\item{paired.end}{a character string indicating how to handle paired-end
reads. If \code{paired.end!="ignore"} then only first reads in proper mapped
pairs will be consider (i.e. in the flag of the read, the bits in the mask
66 must be all ones). If \code{paired.end=="midpoint"} then the midpoint of a
fragment is considered, where \code{mid = fragment_start + int(abs(tlen)/2)},
and where tlen is the template length stored in the bam file. For even tlen,
the given midpoint will be moved of 0.5 basepairs in the 3' direction.
If \code{paired.end=="extend"} then the whole fragment is treated
as a single read.}

\item{verbose}{A logical value indicating whether verbose output is desired}
}
\value{
a \link{list} with the following elements:
	\item{posterior}{a matrix containing posteriors for model components.}
	\item{foldchange}{a matrix containing foldchanges for model components
 (i=2...models) to background component (i=1).}
 \item{fit}{
   Result of the multinomial fit. \code{qstar} is naive estimate of
   background intensity. \code{theta} gives binomial mixture model
   parameters. \code{prior} gives binomial mixture model priors.\code{lnL}
   gives lLn Likelihood trace.
 }
 \item{p.values}{-Log10 P-values for enrichment over background for model
 components (i=2..models).}
}
\description{
This function implements difference calling of two NGS experiments by joint
multinomial modeling. A given number of model components are fit
simultaenously via an efficient Expectation Maximization implementation in
C++. Convergence is achieved by a threshold on the minimum change in model
ln likelihood. The background component is assumed to be the component
with lowest mean and is used to compute two treatments fold change values
and p-values for statistical significance of enrichment.
}
\examples{
\dontrun{
#bam file input
normalize( "H3K4me3.bam", "H3K27me3.bam" )
#count vector input
normalize( H3K4me3.counts, H3K27me3.counts )
#Differences of H3K4me3 and H3K27me3 from bam input with 2 enrichment regimes
normalize( "H3K4me3.bam", "H3K27me3.bam", bin.size=150, models=5 )
}
}

