% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{diffR}
\alias{diffR}
\title{\code{diffR}: Difference calling between \code{treatment} (ChIP-seq 1) and
\code{control} (ChIP-seq 2) for either two given read count vectors and an
optional \link{GenomicRanges} object that defines the original regions
('integer,integer,GenomicRangesOrNULL'), two given bam filepaths and a
genome \code{data.frame} defining chromosomes which will be binned according
to 'countConfig' ('character,character,data.frame') or two given bam
filepaths and a genome \code{character} specifying a UCSC genome identifier
(e.g. "hg19") that is used for retrieval of chromosome annotation which will
be binned according to 'countConfig' ('character,character,character').}
\usage{
diffR(treatment, control, genome, ...)
}
\description{
\code{diffR}: Difference calling between \code{treatment} (ChIP-seq 1) and
\code{control} (ChIP-seq 2) for either two given read count vectors and an
optional \link{GenomicRanges} object that defines the original regions
('integer,integer,GenomicRangesOrNULL'), two given bam filepaths and a
genome \code{data.frame} defining chromosomes which will be binned according
to 'countConfig' ('character,character,data.frame') or two given bam
filepaths and a genome \code{character} specifying a UCSC genome identifier
(e.g. "hg19") that is used for retrieval of chromosome annotation which will
be binned according to 'countConfig' ('character,character,character').
}

