% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dnorm_plot.R
\name{dnorm_plot}
\alias{dnorm_plot}
\title{Plots probability density function of a normal random variable}
\usage{
dnorm_plot(mean, var, by = 0.0025, plot_title = "Normal Density Function")
}
\arguments{
\item{mean}{Mean}

\item{var}{Variance}

\item{by}{Controls granularity}

\item{plot_title}{Controls the title of the plot}
}
\description{
Plots probability density function of a normal random variable
}
\examples{
dnorm_plot(0,1)
}
