% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_mu_posterior.R
\name{draw_mu_posterior}
\alias{draw_mu_posterior}
\title{Draws from the posterior of sigma squared}
\usage{
draw_mu_posterior(mu_n, sigsq_posterior_draw, kappa_n)
}
\arguments{
\item{mu_n}{See function for mu_n}

\item{sigsq_posterior_draw}{Relies on previously drawn sigma^2}
}
\description{
Draws from the posterior of sigma squared
}
