% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_sigsq_posterior.R
\name{draw_sigsq_posterior}
\alias{draw_sigsq_posterior}
\title{Draws from the posterior of sigma squared}
\usage{
draw_sigsq_posterior(n_sim = 10000, nu_n, sigsq_n)
}
\arguments{
\item{n_sim}{Number of draws}

\item{nu_n}{nu_n = nu_0 + n}

\item{sigsq_n}{See function for sigsq_n}
}
\description{
Draws from the posterior of sigma squared
}
