% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{em}
\alias{em}
\title{Deconvolute bivariate count data in multiple enrichment regimes. Bivariate
 data is modeled as a mixture of binomial distributions. Fitting is done
 with Expectation Maximization (EM).}
\usage{
em(r, s, models = 2L, eps = 1e-04, verbose = FALSE, nthreads = 1L)
}
\arguments{
\item{r}{\code{integer}-vector of counts (e.g. control counts in enrichment calls).
If elements are not integers, they will be casted to integers.}

\item{s}{\code{integer}-vector of counts (e.g. treatment counts in enrichment calls).
If elements are not integers, they will be casted to integers.}

\item{models}{\code{integer} specifying number of mixture components which should be
>= 2 (default=2).}

\item{eps}{\code{double} specifying termination criterion for EM fit (default=0.001).}

\item{verbose}{\code{logical} specifying if logging should be performed (default=FALSE).}

\item{nthreads}{\code{integer} specifying number of cores to use (default=1).}
}
\value{
a list with the following items:
       \item{qstar}{naive enrichment ratio \code{s/(r + s)}. Basis for EM fit.}
       \item{theta}{Parametrization for \code{models} binomial distributiions}
       \item{prior}{Mixture proportions for \code{models} binomial distributiions}
       \item{posterior}{Posteriormatrix over all bins for \code{models} binomial distributiions}
       \item{lnL}{log likelihood trace}
}
\description{
Deconvolute bivariate count data in multiple enrichment regimes. Bivariate
 data is modeled as a mixture of binomial distributions. Fitting is done
 with Expectation Maximization (EM).
}

