% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{enrichR}
\alias{enrichR}
\title{Normalize a NGS experiment with given background data.}
\usage{
enrichR(treatment, control, genome, ...)
}
\arguments{
\item{treatment}{A \link{integer} vector of treatment counts or a
\link{character} pointing to the treatment bam file. In the latter case an
\code{treatment}.bai index file should exist in the same folder. Should be
consistent with control.}

\item{control}{A \link{integer} vector of control counts or a
\link{character} pointing to the control bam file. In the latter case an
\code{control}.bai index file should exist in the same folder. Should be
consistent with treatment.}

\item{genome}{Either \code{NULL}, an USCS genome identifier for
\link{GenomeInfoDb::fetchExtendedChromInfoFromUCSC} (only assembled
molecules, circular omitted) or a \link{data.frame} consisting of two
columns(1st column=chromosome names, 2nd column=lengths). Chromosome names
given should be present in the bam file header. If \code{NULL}, chromosome
names will be read from treatment bamheader. Please be aware that bamheader
might contain irregular contigs and chrM which influence the diffR fit.
(DEFAUlT=NULL).}

\item{bin.size}{Width of genomic bins in bp.}

\item{models}{Number of model components.}

\item{eps}{Threshold for EM convergence.}

\item{procs}{Number of threads to use}

\item{mapqual}{discard reads with mapping quality strictly lower than this
parameter. The value 0 ensures that no read will be discarded, the value 254
that only reads with the highest possible mapping quality will be considered.}

\item{shift}{shift the read position by a user-defined number of basepairs.
This can be handy in the analysis of chip-seq data.}

\item{paired.end}{a character string indicating how to handle paired-end
reads. If \code{paired.end!="ignore"} then only first reads in proper mapped
pairs will be consider (i.e. in the flag of the read, the bits in the mask
66 must be all ones). If \code{paired.end=="midpoint"} then the midpoint of a
fragment is considered, where \code{mid = fragment_start + int(abs(tlen)/2)},
and where tlen is the template length stored in the bam file. For even tlen,
the given midpoint will be moved of 0.5 basepairs in the 3' direction.
If \code{paired.end=="extend"} then the whole fragment is treated
as a single read.}

\item{verbose}{A logical value indicating whether verbose output is desired}
}
\value{
a \link{list} with the following elements:
 \item{posterior}{a matrix containing posteriors for model components.}
 \item{fit}{
   Result of the multinomial fit. \code{qstar} is naive estimate of
   background intensity. \code{theta} gives binomial mixture model
   parameters. \code{prior} gives binomial mixture model priors.\code{lnL}
   gives lLn Likelihood trace.
 }
}
\description{
This function implements the normalization of a treatment track with its
given control track by joint multinomial modeling. A given number of model
components are fit simultaenously via an efficient Expectation
Maximization implementation in C++. Convergence is achieved by a
threshold on the minimum change in model ln likelihood. The background
component is assumed to be the component with lowest mean and is used
to compute treatment fold change and P-Values for statistical significance
of enrichment.
}
\examples{

}

