% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getEnrichment}
\alias{getEnrichment}
\title{Get normalized enrichment from a diffR fit}
\usage{
getEnrichment(posteriors, r, s, k = 1L)
}
\arguments{
\item{posteriors}{posterior matrix as computed by diffR}

\item{r}{vector of counts in control}

\item{s}{vector of counts in treatment}

\item{k}{column index of background component in posteriors (DEFAULT=1)}
}
\value{
a numeric with enrichment values in log space
}
\description{
Get normalized enrichment from a diffR fit
}

