% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mapToOriginal}
\alias{mapToOriginal}
\title{Retrieve original vector of values with a supplied map}
\usage{
mapToOriginal(vec, m2u)
}
\arguments{
\item{vec}{The \code{numeric()}-vector to be mapped back.}

\item{map}{A map computed by \code{\link{mapToUniquePairs}}.}
}
\value{
a \code{numeric()}-vector of mapped back values.
}
\description{
When computing only on unique values, it is often desired to retrieve the
original ordering for output. This function does this by creating a new
numeric().
}
\seealso{
\code{\link{mapToUniquePairs}} for generation of map
}

