% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mapToUniquePairs}
\alias{mapToUniquePairs}
\title{Group unique tupels from two integer vectors r and s}
\usage{
mapToUniquePairs(r, s)
}
\arguments{
\item{r}{An \code{integer()}-vector If elements are not integers, they
will be casted to integers.}

\item{s}{An \code{integer()}-vector. If elements are not integers, they 
will be casted to integers.}
}
\value{
a list with the following items:
       \item{values}{unique and sorted values of \code{r} and \code{s}}
       \item{map}{a vector such that 
                  \code{cbind(r,s)[i,] = values[,map[i]]} for every i}
}
\description{
Group unique tupels from two integer vectors r and s
}

