% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarginPlot.R
\name{MarginPlot}
\alias{MarginPlot}
\title{A margin plot show the VOI density distribution of a individual measure by groups}
\usage{
MarginPlot(dfMargin, group = NULL, ylabel = "Expected Probability",
  xlabel = "Level of voi")
}
\arguments{
\item{dfMargin}{A dateframe with margin value from \link[normr]{AvgMargin} dataframe.}
}
\value{
A margin plot with specified value
}
\description{
It is written after getting the average predicted value from \link[normr]{AvgMargin}
}
\keyword{plot}
