% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu_n.R
\name{mu_n}
\alias{mu_n}
\title{Computes mu_n from prior parameters}
\usage{
mu_n(kappa_0, mu_0, n, y_bar)
}
\arguments{
\item{kappa_0}{number of prior observations}

\item{mu_0}{mean of prior observations}

\item{n}{number of new observations}

\item{y_bar}{mean of new observations}
}
\description{
Computes mu_n from prior parameters
}
