% Generated by roxygen2 (4.0.2): do not edit by hand
\name{normalize}
\alias{normalize}
\title{Normalize a NGS experiment with given background data.}
\usage{
normalize(treatment, control, genome, bin.size = 300, models = 2,
  eps = 0.001, procs = 1, mapqual = 20, shift = 0, paired.end = F,
  paired.end.midpoint = T, verbose = T)
}
\arguments{
\item{treatment}{A \link{numeric} vector of treatment counts or a
\link{character} pointing to the treatment bam file. In the latter case an
\code{treatment}.bai index file should exist in the same folder. Should be
consistent with control.}

\item{control}{A \link{numeric} vector of control counts or a
\link{character} pointing to the control bam file. In the latter case an
\code{control}.bai index file should exist in the same folder. Should be
consistent with treatment.}

\item{genome}{A \link{data.frame} consisting of two columns. First column
gives sequence names. Second column gives chromosome lengths. This
information can be retrieved via the UCSC chromSizes table.}

\item{bin.size}{Width of genomic bins in bp.}

\item{models}{Number of model components.}

\item{eps}{Threshold for EM convergence.}

\item{procs}{Number of threads to use in parallel::mclapply()}

\item{mapqual}{discard reads with mapping quality strictly lower than this
parameter. The value 0 ensures that no read will be discarded, the value 254
that only reads with the highest possible mapping quality will be considered.}

\item{shift}{shift the read position by a user-defined number of basepairs.
This can be handy in the analysis of chip-seq data.}

\item{paired.end}{A logical value indicating whether the bampath contains
paired-end sequencing output. In this case, only first reads in proper mapped
pairs are considered (FLAG 66).}

\item{paired.end.midpoint}{A logical value indicating whether the fragment
middle points of each fragment should be counted. Therefore it uses the tlen
information from the given bam file (MidPoint = fragment_start + int(
abs(tlen) / 2) )). For even tlen, the given midpoint is the round half down
real midpoint.}

\item{verbose}{A logical value indicating whether verbose output is desired}
}
\value{
a \link{list} with the following elements:
	\item{posterior}{a matrix containing posteriors for model components.}
	\item{foldchange}{a matrix containing foldchanges for model components
 (i=2...models) to background component (i=1).}
 \item{fit}{
   Result of the multinomial fit. \code{qstar} is naive estimate of
   background intensity. \code{theta} gives binomial mixture model
   parameters. \code{prior} gives binomial mixture model priors.\code{lnL}
   gives lLn Likelihood trace.
 }
 \item{p.values}{-Log10 P-values for enrichment over background for model
 components (i=2..models).}
}
\description{
This function implements the normalization of a treatment track with its
given control track by joint multinomial modeling. A given number of model
components are fit simultaenously via an efficient Expectation
Maximization implementation in C++. Convergence is achieved by a
threshold on the minimum change in model ln likelihood. The background
component is assumed to be the component with lowest mean and is used
to compute treatment fold change and P-Values for statistical significance
of enrichment.
}

