% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wrappers.R
\docType{methods}
\name{diffR,character,character-method}
\alias{diffR,character,character-method}
\alias{diffR,numeric,numeric-method}
\alias{differenceCall}
\title{\code{diffR}: Does a difference call for two conditions. See details.}
\usage{
\S4method{diffR}{character,character}(condition1, condition2, genome = NULL,
  bin.size = 300, models = 2, eps = 0.001, p.values = T, procs = 1,
  mapqual = 20, shift = 0, paired.end = "ignore", verbose = T)

\S4method{diffR}{numeric,numeric}(condition1, condition2, genome = NULL,
  bin.size = 300, models = 2, eps = 0.001, p.values = T, procs = 1,
  mapqual = 20, shift = 0, paired.end = "ignore", verbose = T)
}
\description{
\code{diffR}: Does a difference call for two conditions. See details.

\code{diffR}:
}

