% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{normr_core}
\alias{normr_core}
\title{Deconvolute bivariate count data in multiple enrichment regimes. Bivariate
data is modeled as a mixture of binomial distributions. Fitting is done
with Expectation Maximization (EM) on data points were \code{r > 0 & s > 0}.}
\usage{
normr_core(r, s, models = 2L, eps = 1e-04, iterations = 5L, bgIdx = 0L,
  diffCall = FALSE, verbose = FALSE, nthreads = 1L)
}
\arguments{
\item{r}{\code{integer}-vector of counts (e.g. control counts in enrichment 
calls). If elements are not integers, they will be casted to integers.}

\item{s}{\code{integer}-vector of counts (e.g. treatment counts in 
enrichment calls). If elements are not integers, they will be casted to 
integers.}

\item{models}{\code{integer} specifying number of mixture components which 
should be >= 2 (default=2).}

\item{eps}{\code{double} specifying termination criterion for EM fit 
(default=0.001).}

\item{iterations}{\code{integer} specifying the number of individual EM runs
with differential initial parameters to be done. Adjust this argument to
find global maxima (default=5).}

\item{bgIdx}{\code{integer} giving the index of the background component. In
enrichment and regime calls this should be 0. In difference calls, this value
can be > 0 (default=0).}

\item{diffCall}{\code{logical} specifying if difference calling is done such
that a two-sided significance test will be conducted (default=FALSE).}

\item{verbose}{\code{logical} specifying if logging should be performed (default=FALSE).}

\item{nthreads}{\code{integer} specifying number of cores to use (default=1).}
}
\value{
a list with the following items:
 \item{qstar}{naive enrichment ratio \code{s/(r + s)}. Basis for EM fit.}
 \item{map}{a map of unique (r,s) values. See
  \code{\link{map2uniquePairs()}}}
 \item{lntheta}{ln parametrization of mixture binomials}
 \item{lnprior}{ln mixture proportions of mixture binomials}
 \item{lnL}{log likelihood trace of EM}
 \item{lnposterior}{ln posteriors for unique (r,s) according to map}
 \item{lnenrichment}{ln enrichment for unique (r,s) according to map}
 \item{lnpvals}{ln P-values for mixture component \code{bgIdx} for each 
  unique (r,s)}
 \item{filtered}{unique (r,s) tupels passing the T filter with \code{eps}}
}
\description{
In a first step, a map of unique non-zero (r,s) values is generated. This 
allows for faster runtime. Second, EM is run with the given number of
components on these reduced data representation. If \code{iterations > 1},
the fit with the highest likelihood is selected. In a third step, a map for
all (r,s) values is generated. Based on this map the complete posterior
matrix, P-values and enrichment is calculated. The returned values are
holding results exclusively for unique (r,s) values and go directly into a 
\code{\link{NormRFit-class}} object.
}

