% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pie.plot.R
\name{pie.plot}
\alias{pie.plot}
\title{pie.plot a var by group in df}
\usage{
pie.plot(data, group, var)
}
\description{
\code{pie.plot} take the categorical factor variables output a ggplot pie chart by category.
}
\details{
This function is built upon ggplot2 pie chart function.
}
