% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_monte_sim.R
\name{plot_monte_sim}
\alias{plot_monte_sim}
\title{Plots the monte carlo draws}
\usage{
plot_monte_sim(sig_sim, mu_sim)
}
\arguments{
\item{sig_sim}{Sigma^2 draws}

\item{mu_sim}{mu draws}
}
\description{
Plots the monte carlo draws
}
