% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_posterior_density_estimate.R
\name{plot_posterior_density_estimate}
\alias{plot_posterior_density_estimate}
\title{Plots univariate density estimate}
\usage{
plot_posterior_density_estimate(x_sim, plot_title = "Posterior density")
}
\arguments{
\item{x_sim}{Draws from posterior density}

\item{plot_title}{Gives plot a title}
}
\description{
Plots univariate density estimate
}
