% Generated by roxygen2 (4.0.2): do not edit by hand
\name{processByChromosome}
\alias{processByChromosome}
\title{Process a list of bam files with a number of processes.}
\usage{
processByChromosome(bam.files, gr, procs, bamsignals.function, mapqual,
  shift = 0, paired.end = F, paired.end.midpoint = T, verbose = F)
}
\arguments{
\item{bam.files}{A list of strings specifying file system locations for bam
files to count.}

\item{gr}{A\link{GenomicRanges}-object giving the genomic intervals to count
in.}

\item{procs}{Number of threads to use in parallel::mclapply()}

\item{bamsignals.function}{The corresponding counting function in bamsignals
to call.}

\item{mapqual}{discard reads with mapping quality strictly lower than this
parameter. The value 0 ensures that no read will be discarded, the value 254
that only reads with the highest possible mapping quality will be considered.}

\item{shift}{shift the read position by a user-defined number of basepairs.
This can be handy in the analysis of chip-seq data.}

\item{paired.end}{A logical value indicating whether the bampath contains
paired-end sequencing output. In this case, only first reads in proper mapped
pairs are considered (FLAG 66).}

\item{paired.end.midpoint}{A logical value indicating whether the fragment
middle points of each fragment should be counted. Therefore it uses the tlen
information from the given bam file (MidPoint = fragment_start + int(
abs(tlen) / 2) )). For even tlen, the given midpoint is the round half down
real midpoint.}

\item{verbose}{A logical value indicating whether verbose output is desired}
}
\value{
a list of length(bam.files) with bamsignals.function results
}
\description{
Process a list of bam files with a number of processes.
}

