% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomRows.R
\name{randomRows}
\alias{randomRows}
\title{Generate random rows from a dataframe}
\usage{
randomRows(df, n)
}
\arguments{
\item{n}{number of random rows}
}
\description{
Generate random rows from a dataframe
}
