% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relabelSpssVariable.R
\name{relabelSpssVariable}
\alias{relabelSpssVariable}
\title{reapply labels to SPSS dataset}
\usage{
relabelSpssVariable(x)
}
\arguments{
\item{x}{spss dataset}
}
\description{
reapply labels to SPSS dataset
}
