% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgeDensityIdv.R
\name{ridgeDensityIdv}
\alias{ridgeDensityIdv}
\title{A ridge plot show the VOI density distribution of a individual measure by groups}
\usage{
ridgeDensityIdv(data, dfBoxplot, dv,
  xlabel = "Proportion of Community Whom Respondent Believes Uses a Toilet")
}
\arguments{
\item{data}{A dateframe with var "psutype" and "psuname".}

\item{dfBoxplot}{A dataframe contain value calcualted from \link[normr]{densityQuantile}..}

\item{dv}{Variable of interest to be calculated.}
}
\value{
A plot show the VOI density ridge and boxplot with specified value
}
\description{
\code{ridgeDensityIdv} A ridge plot show the VOI density distribution by groups
It is written for supplying value to be specified in the boxplot.
The boxplot value is calculated by the \link[normr]{densityQuantile}.
}
\keyword{plot}
