% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgeDensityPsu.R
\name{ridgeDensityPsu}
\alias{ridgeDensityPsu}
\title{A ridge plot show the VOI density distribution of a PSU level measure by groups}
\usage{
ridgeDensityPsu(data, dfBoxplot, dv,
  xlabel = "Proportion of Sampling Unit Who Used a Toilet For Last Defecation",
  captiontext = "Figure 2. Kernel density estimates and box plots with individual data points of sampling unit level prevalence \\nof toilet use for last defecation. Data disaggregated by geography type.")
}
\arguments{
\item{data}{A dateframe with var "psutype" and "psuname".}

\item{dfBoxplot}{A dataframe contain value calcualted from \link[normr]{densityQuantile}..}

\item{dv}{Variable of interest to be calculated.}
}
\value{
A plot show the VOI density ridge and boxplot with specified value
}
\description{
\code{ridgeDensityPsu} A ridge plot show the VOI density distribution by groups
It is written for supplying value to be specified in the boxplot.
The boxplot value is calculated by the \link[normr]{densityQuantile}.
}
\keyword{plot}
