% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigsq_n.R
\name{sigsq_n}
\alias{sigsq_n}
\title{Calculates sigsn_n from prior parameters}
\usage{
sigsq_n(nu_0, sigsq_0, n, ssq, kappa_0, y_bar, mu_0)
}
\arguments{
\item{nu_0}{prior sample size}

\item{sigsq_0}{prior sample variance}

\item{n}{new sample size}

\item{ssq}{new sample variance}

\item{kappa_0}{prior sample size of means}

\item{y_bar}{current sample mean}

\item{mu_0}{prior sample mean}
}
\description{
Calculates sigsn_n from prior parameters
}
