% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormRFit.R
\docType{methods}
\name{writeBed,NormRFit-method}
\alias{writeBed}
\alias{writeBed,NormRFit-method}
\title{describeIn NormRFit Writes a bed file of regions called significant for}
\usage{
\S4method{writeBed}{NormRFit}(obj, filename = NULL, pval = 0.05,
  gr = NULL)
}
\arguments{
\item{obj}{A NormRFit object}

\item{filename}{A file to write to.}

\item{pval}{A p-value threshold.}

\item{gr}{A GenomicRanges object representing the analyzed data. Usually, 
ranges are stored inside \code{obj}. Use this if you want to change
chromosome identifiers for example. Note that the width of the bins should be
same to data and fit stored in \code{obj}.}
}
\description{
describeIn NormRFit Writes a bed file of regions called significant for
}

