% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormRFit.R
\docType{methods}
\name{writeEnrichment,NormRFit-method}
\alias{writeEnrichment}
\alias{writeEnrichment,NormRFit-method}
\title{describeIn NormRFit Writes the calculated enrichment of a NormRFit to disk.}
\usage{
\S4method{writeEnrichment}{NormRFit}(obj, filename = NULL, gr = NULL)
}
\arguments{
\item{obj}{A NormRFit object}

\item{filename}{A file to write to.}

\item{gr}{A GenomicRanges object representing the analyzed data. Usually,
ranges are stored inside \code{obj}. Use this if you want to change
chromosome identifiers for example. Note that the width of the bins should be
same to data and fit stored in \code{obj}.}
}
\description{
describeIn NormRFit Writes the calculated enrichment of a NormRFit to disk.
}

