% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{computeEnrichment}
\alias{computeEnrichment}
\title{Get normalized enrichment from a diffR fit}
\usage{
computeEnrichment(r, s, lnPost, theta, F = 1L, B = 0L, diffCall = FALSE,
  nthreads = 1L)
}
\arguments{
\item{r}{vector of counts in control/condition1}

\item{s}{vector of counts in treatment/condition2}

\item{lnPost}{log-posterior matrix as computed by a normR routine}

\item{theta}{numeric vector of binomial model parameters}

\item{F}{column index of foreground component in posteriors (DEFAULT=1)}

\item{B}{column index of background component in posteriors (DEFAULT=0)}

\item{diffCall}{logical indicating if a difference call was performed. In
that case standardization is performed dependent on the sign of the fold
change. (DEFAULT=FALSE)}
}
\value{
a numeric with enrichment values in log space
}
\description{
Get normalized enrichment from a diffR fit
}

