% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{diffr}
\alias{diffr}
\alias{diffr-package}
\title{Normalization and difference calling for Next Generation Sequencing (NGS)
experiments via joint multinomial modeling}
\description{
Functions for normalization and difference calling in NGS experiment setups.
A binomial mixture model with a given number of components is fit and used
for identifying enriched or depleted regions in two given data tracks.
Log-space multinomial model is fit by Expectation maximization in C/C++.
}
\author{
Johannes Helmuth \email{helmuth@molgen.mpg.de}
}

