% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{enrichR}
\alias{enrichR}
\title{Normalize a NGS experiment with given background data.}
\usage{
enrichR(treatment, control, genome, ...)
}
\arguments{
\item{treatment}{A \link{integer} vector of treatment counts or a
\link{character} pointing to the treatment bam file. In the latter case an
\code{treatment}.bai index file should exist in the same folder. Should be
consistent with control.}

\item{control}{A \link{integer} vector of control counts or a
\link{character} pointing to the control bam file. In the latter case an
\code{control}.bai index file should exist in the same folder. Should be
consistent with treatment.}

\item{genome}{Either \code{NULL}, an USCS genome identifier for
\link{GenomeInfoDb::fetchExtendedChromInfoFromUCSC} (only assembled
molecules, circular omitted) or a \link{data.frame} consisting of two
columns(1st column=chromosome names, 2nd column=lengths). Chromosome names
given should be present in the bam file header. If \code{NULL}, chromosome
names will be read from treatment bamheader. Please be aware that bamheader
might contain irregular contigs and chrM which influence the diffR fit.
(DEFAUlT=NULL).}

\item{bin.size}{Width of genomic bins in bp.}

\item{models}{Number of model components.}

\item{eps}{Threshold for EM convergence.}

\item{procs}{Number of threads to use}

\item{countConfig}{A BamCountConfig object specifying bam counting parameters
for read count retrieval.}

\item{filename}{A filename to write data to.}

\item{threshold}{A threshold on q-values. Set to 1 to write all regions.}

\item{verbose}{A logical value indicating whether verbose output is desired}
}
\value{
a \link{NormRFit} object
}
\description{
This function implements the normalization of a treatment track with its
given control track by joint multinomial modeling. A given number of model
components are fit simultaenously via an efficient Expectation
Maximization implementation in C++. Convergence is achieved by a
threshold on the minimum change in model ln likelihood. The background
component is assumed to be the component with lowest mean and is used
to compute treatment fold change and P-Values for statistical significance
of enrichment.
}
\examples{

}
\seealso{
BamCountConfig-class

NormRFit-class
}

