% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mapToUniquePairs}
\alias{mapToUniquePairs}
\title{Group unique pairs from two vectors}
\usage{
mapToUniquePairs(r, s)
}
\arguments{
\item{r}{first vector of integers. If elements are not integers, they will be
casted to integers.}

\item{s}{second vector of integers. If elements are not integers, they will be
casted to integers.}
}
\value{
a list with the following items:
       \item{values}{unique and sorted values of \code{r} and \code{s}}
       \item{map}{a vector such that \code{cbind(r,s)[i,] = values[,map[i]]} for every i}
}
\description{
Group unique pairs from two vectors
}

