% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{enrichR,integer,integer,GenomicRangesOrNULL-method}
\alias{computeP}
\alias{enrichR}
\alias{enrichR,character,character,character-method}
\alias{enrichR,character,character,data.frame-method}
\alias{enrichR,integer,integer,GenomicRangesOrNULL-method}
\alias{exportR}
\alias{exportR,NormRFit,character,character-method}
\alias{recomputeP,NormRFit,integer-method}
\title{\code{enrichR}: Do enrichment calling between treatment (ChIP-seq) and
control (Input) for two given bam files and a genome data.frame}
\usage{
\S4method{enrichR}{integer,integer,GenomicRangesOrNULL}(treatment, control,
  genome = NULL, eps = 1e-05, iterations = 10, procs = 1L,
  verbose = TRUE)

\S4method{enrichR}{character,character,data.frame}(treatment, control, genome,
  countConfig = countConfigSingleEnd(), eps = 1e-05, iterations = 10,
  procs = 1L, verbose = TRUE)

\S4method{enrichR}{character,character,character}(treatment, control,
  genome = "", countConfig = countConfigSingleEnd(), eps = 1e-05,
  iterations = 10, procs = 1L, verbose = TRUE)

\S4method{recomputeP}{NormRFit,integer}(fit, B)

\S4method{exportR}{NormRFit,character,character}(fit, filename,
  format = c("bed", "bigWig"), threshold = 0.05)
}
\description{
\code{enrichR}: Do enrichment calling between treatment (ChIP-seq) and
control (Input) for two given bam files and a genome data.frame

\code{enrichR}: Do enrichment calling between treatment (ChIP-seq) and
control (Input) for two given bam files and a genome data.frame

\code{enrichR}: Do enrichment calling between treatment (ChIP-seq) and
control (Input) for two given bam files and a genome data.frame

\code{recomputeP}: Recompute P-values relative to provided background
component B.

\code{exportR}: Export results of a normR fit. The provided filetype
specifies if significant regions (\code{"bed"}) or calculated enrichment
(\code{c("bigWig", "bedGraph")}) is exported
}

