% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{normr-methods}
\alias{diffR}
\alias{diffR,character,character,GenomicRanges-method}
\alias{diffR,character,character,character-method}
\alias{diffR,character,character,data.frame-method}
\alias{diffR,integer,integer,GenomicRanges-method}
\alias{enrichR}
\alias{enrichR,character,character,GenomicRanges-method}
\alias{enrichR,character,character,character-method}
\alias{enrichR,character,character,data.frame-method}
\alias{enrichR,integer,integer,GenomicRanges-method}
\alias{normr-methods}
\alias{regimeR}
\alias{regimeR,character,character,GenomicRanges,integer-method}
\alias{regimeR,character,character,character,integer-method}
\alias{regimeR,character,character,data.frame,integer-method}
\alias{regimeR,integer,integer,GenomicRanges,integer-method}
\title{Enrichment, Difference and Regime Calling in ChIP-seq data.}
\usage{
enrichR(treatment, control, genome, ...)

\S4method{enrichR}{integer,integer,GenomicRanges}(treatment, control, genome,
  fdr = 0.05, eps = 1e-05, iterations = 10, procs = 1L,
  verbose = TRUE)

\S4method{enrichR}{character,character,GenomicRanges}(treatment, control,
  genome, countConfig = countConfigSingleEnd(), fdr = 0.05, eps = 1e-05,
  iterations = 10, procs = 1L, verbose = TRUE)

\S4method{enrichR}{character,character,data.frame}(treatment, control, genome,
  countConfig = countConfigSingleEnd(), fdr = 0.05, eps = 1e-05,
  iterations = 10, procs = 1L, verbose = TRUE)

\S4method{enrichR}{character,character,character}(treatment, control, genome,
  countConfig = countConfigSingleEnd(), fdr = 0.05, eps = 1e-05,
  iterations = 10, procs = 1L, verbose = TRUE)

diffR(treatment, control, genome, ...)

\S4method{diffR}{integer,integer,GenomicRanges}(treatment, control, genome,
  fdr = 0.05, eps = 1e-05, iterations = 10, procs = 1L,
  verbose = TRUE)

\S4method{diffR}{character,character,GenomicRanges}(treatment, control, genome,
  countConfig = countConfigSingleEnd(), fdr = 0.05, eps = 1e-05,
  iterations = 10, procs = 1L, verbose = TRUE)

\S4method{diffR}{character,character,data.frame}(treatment, control, genome,
  countConfig = countConfigSingleEnd(), fdr = 0.05, eps = 1e-05,
  iterations = 10, procs = 1L, verbose = TRUE)

\S4method{diffR}{character,character,character}(treatment, control,
  genome = "", countConfig = countConfigSingleEnd(), fdr = 0.05,
  eps = 1e-05, iterations = 10, procs = 1L, verbose = TRUE)

regimeR(treatment, control, genome, models, ...)

\S4method{regimeR}{integer,integer,GenomicRanges,integer}(treatment, control,
  genome, models = 3L, fdr = 0.05, eps = 1e-05, iterations = 10,
  procs = 1L, verbose = TRUE)

\S4method{regimeR}{character,character,GenomicRanges,integer}(treatment,
  control, genome, models = 3L, countConfig = countConfigSingleEnd(),
  fdr = 0.05, eps = 1e-05, iterations = 10, procs = 1L,
  verbose = TRUE)

\S4method{regimeR}{character,character,data.frame,integer}(treatment, control,
  genome, models = 3L, countConfig = countConfigSingleEnd(), fdr = 0.05,
  eps = 1e-05, iterations = 10, procs = 1L, verbose = TRUE)

\S4method{regimeR}{character,character,character,integer}(treatment, control,
  genome = "", models = 3L, countConfig = countConfigSingleEnd(),
  fdr = 0.05, eps = 1e-05, iterations = 10, procs = 1L,
  verbose = TRUE)
}
\arguments{
\item{treatment}{A \link{integer} vector of treatment counts or a
\link{character} pointing to the treatment bam file. In the latter case an
\code{treatment}.bai index file should exist in the same folder. Should be
consistent with control.}

\item{control}{A \link{integer} vector of control counts or a
\link{character} pointing to the control bam file. In the latter case an
\code{control}.bai index file should exist in the same folder. Should be
consistent with treatment.}

\item{genome}{Either \code{NULL}, an USCS genome identifier for
\code{GenomeInfoDb::fetchExtendedChromInfoFromUCSC} (only assembled
molecules, circular omitted) or a \link{data.frame} consisting of two
columns(1st column=chromosome names, 2nd column=lengths). Chromosome names
given should be present in the bam file header. If \code{NULL}, chromosome
names will be read from treatment bamheader. Please be aware that bamheader
might contain irregular contigs and chrM which influence the diffR fit.
(DEFAUlT=NULL).}

\item{fdr}{The FDR threshold for class definition. See \link{NormRFit}.}

\item{eps}{Threshold for EM convergence.}

\item{procs}{Number of threads to use}

\item{verbose}{A logical value indicating whether verbose output is desired}

\item{countConfig}{A BamCountConfig object specifying bam counting parameters
for read count retrieval. See \link{BamCountConfig}.}

\item{models}{Number of model components for \code{regimeR()}. Should be an
\code{integer() >= 3}. (DEFAULT=3)}
}
\value{
a \link{NormRFit} object
}
\description{
A correct background estimation is crucial for calling enrichment in ChIP-seq
data.  These functions implement the mixture modeling for two given ChIP-seq
tracks by joint multinomial modeling. A predefined number of model components
is fit simultaneously to binned read count data via an efficient Expectation
Maximization implementation in C++. Convergence is achieved by a threshold on
the minimum change in model loglikelihood. After the model is fit, every bin
is tested for significance against the fitted background component. Moreover,
a standardized enrichment for each bin is calculated based on the fitted
background component.

\code{enrichR}: Enrichment calling between \code{treatment} (ChIP-seq) and
\code{control} (Input) for either two given read count vectors and an
optional \link{GenomicRanges} object that defines the original regions
('integer,integer,GenomicRangesOrNULL'), two given bam filepaths and a
genome \code{data.frame} defining chromosomes which will be binned according
to 'countConfig' ('character,character,data.frame') or two given bam
filepaths and a genome \code{character} specifying a UCSC genome identifier
(e.g. "hg19") that is used for retrieval of chromosome annotation which will
be binned according to 'countConfig' ('character,character,character').

\code{diffR}: Difference calling between \code{treatment} (ChIP-seq 1) and
\code{control} (ChIP-seq 2) for either two given read count vectors and an
optional \link{GenomicRanges} object that defines the original regions
('integer,integer,GenomicRangesOrNULL'), two given bam filepaths and a
genome \code{data.frame} defining chromosomes which will be binned according
to 'countConfig' ('character,character,data.frame') or two given bam
filepaths and a genome \code{character} specifying a UCSC genome identifier
(e.g. "hg19") that is used for retrieval of chromosome annotation which will
be binned according to 'countConfig' ('character,character,character').

\code{regimeR}: Enrichment regime calling between \code{treatment}
(ChIP-seq 1) and \code{control} (ChIP-seq 2) with with \code{models} number
of models for either two given read count vectors and an optional
\link{GenomicRanges} object that defines the original regions
('integer,integer,GenomicRangesOrNULL'), two given bam filepaths and a
genome \code{data.frame} defining chromosomes which will be binned according
to 'countConfig' ('character,character,data.frame') or two given bam
filepaths and a genome \code{character} specifying a UCSC genome identifier
(e.g. "hg19") that is used for retrieval of chromosome annotation which will
be binned according to 'countConfig' ('character,character,character').
}
\examples{
inputfile <- system.file("extdata", "K562_Input.bam", package="normr")
chipfiles <- system.file("extdata", "K562_H3K27me3.bam", package="normr")
gr <- GRanges("chr1", IRanges(22500000, 25000000))
fit <- enrichR(chipfile, inputfile, genome.df, countConfigPairedEnd(), verbose=F)
summary(fit)
}
\seealso{
BamCountConfig-class

NormRFit-class
}

