% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{enrichR,integer,integer,GenomicRanges-method}
\alias{enrichR}
\alias{enrichR,character,character,character-method}
\alias{enrichR,character,character,data.frame-method}
\alias{enrichR,integer,integer,GenomicRanges-method}
\title{\code{enrichR}: Do enrichment calling between treatment (ChIP-seq) and
control (Input) for two given bam files and a genome data.frame}
\usage{
\S4method{enrichR}{integer,integer,GenomicRanges}(treatment, control,
  genome = NULL, eps = 1e-05, procs = 1L, verbose = TRUE)

\S4method{enrichR}{character,character,data.frame}(treatment, control, genome,
  countConfig = countConfigSingleEnd(), eps = 1e-05, procs = 1L,
  verbose = TRUE)

\S4method{enrichR}{character,character,character}(treatment, control, genome,
  countConfig = countConfigSingleEnd(), eps = 1e-05, procs = 1L,
  verbose = TRUE)
}
\description{
\code{enrichR}: Do enrichment calling between treatment (ChIP-seq) and
control (Input) for two given bam files and a genome data.frame

\code{enrichR}: Do enrichment calling between treatment (ChIP-seq) and
control (Input) for two given bam files and a genome data.frame

\code{enrichR}: Do enrichment calling between treatment (ChIP-seq) and
control (Input) for two given bam files and a genome data.frame
}

