% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{package}
\name{normr}
\alias{normR}
\alias{normr}
\alias{normr-package}
\title{Normalization and difference calling in ChIP-seq data}
\description{
Robust normalization and difference calling procedures for ChIP-seq and
alike data. Count vectors can be obtained from experiment bam files in two
experiments. Counts are jointly modeled as a multinomial sampling trial.
A binomial mixture model with a given number of components is fit and used
for identifying enriched or depleted regions in two given data tracks.
Log-space multinomial model is fit by Expectation maximization in C++.
}
\author{
Johannes Helmuth \email{helmuth@molgen.mpg.de}
}
\seealso{
\code{\link{normr-methods}} for available functions
}

