% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormRFit.R
\docType{class}
\name{NormRFit-class}
\alias{NormRFit}
\alias{NormRFit-class}
\alias{getClasses}
\alias{getClasses,NormRFit-method}
\alias{getCounts}
\alias{getCounts,NormRFit-method}
\alias{getEnrichment}
\alias{getEnrichment,NormRFit-method}
\alias{getPosteriors}
\alias{getPosteriors,NormRFit-method}
\alias{getPvalues}
\alias{getPvalues,NormRFit-method}
\alias{getQvalues}
\alias{getQvalues,NormRFit-method}
\alias{getRanges}
\alias{getRanges,NormRFit-method}
\alias{length}
\alias{length,NormRFit-method}
\alias{summary,NormRFit-method}
\title{Container for a NormR fit}
\usage{
\S4method{length}{NormRFit}(x)

\S4method{summary}{NormRFit}(object, print = T, digits = 3, ...)

\S4method{getCounts}{NormRFit}(obj)

\S4method{getRanges}{NormRFit}(obj)

\S4method{getPosteriors}{NormRFit}(obj)

\S4method{getEnrichment}{NormRFit}(obj)

\S4method{getPvalues}{NormRFit}(obj, filtered = F)

\S4method{getQvalues}{NormRFit}(obj)

\S4method{getClasses}{NormRFit}(obj)
}
\arguments{
\item{object}{A NormRFit objectect}

\item{...}{Not used}

\item{obj}{A NormRFit object.}

\item{i}{\code{integer()}-index for subsetting.}

\item{B}{The index of the background component.}
}
\value{
return values are described in the methods section.
}
\description{
This S4 class wraps a NormR fit containing counts, fit configuration and
results of the fit.
}
\details{
This class provides printing, summary and accessor methods for retrieving
data from the object. Internally this class uses a map structure which saves
memory, such that accessing data through \code{slot()} gives undesired
output. Please use accessor methods. A NormRFit object is read-only, i.e. it
can not be modified.
}
\section{Methods (by generic)}{
\itemize{
\item \code{length}: Number of regions analyzed.

\item \code{summary}: Prints a summary of the NormRFit object.

\item \code{getCounts}: Retrieve read count data.

\item \code{getRanges}: Retrieve read count data.

\item \code{getPosteriors}: Retrieve computed posteriors.

\item \code{getEnrichment}: Retrieve NormR calculated enrichment.

\item \code{getPvalues}: Retrieve computed P-values.

\item \code{getQvalues}: Retrieve computed Q-values. See Bioconductor package
\link{qvalue}.

\item \code{getClasses}: Retrieve classes for each bin, i.e. enriched vs
non-enriched (enrichR), differential enrichment (diffR) and enrichment
regimes (regimeR).
}}
\section{Slots}{

\describe{
\item{\code{type}}{A \code{character()} representing the type of fit ("enrichR",
"diffR", "regimeR")}

\item{\code{n}}{Number of regions analyzed in total.}

\item{\code{ranges}}{A \code{GenomicRanges}-object specifying the analysed regions.}

\item{\code{k}}{An \code{integer()} giving the number of binomial mixture
components to be fit to the data.}

\item{\code{B}}{The index of the background component.}

\item{\code{map}}{A \code{integer}-vector to map back \code{counts},
\code{lnposteriors}, \code{lnenrichment}, \code{lnpvals} and \code{lnqvals}.
See also \code{map2uniquePairs()} for map generation.}

\item{\code{counts}}{A \code{list()} of length 2 containing unique count pairs. Use
accessor \code{\link{getCounts}} to retrieve original count matrix.}

\item{\code{names}}{The corresponding names for counts.}

\item{\code{thetaStar}}{A \code{numeric()} representing a naive background
estimation, i.e. \code{sum(getCounts(obj)[2,])/sum(getCounts(obj))}}

\item{\code{theta}}{A \code{k}-length \code{numeric()}-vector representing
NormR-fitted parametrization of binomial mixture components.}

\item{\code{mixtures}}{A \code{k}-length \code{numeric()}-vector representing
NormR-fitted mixture proportions of binomial mixture
components.}

\item{\code{lnL}}{A \code{numeric()}-vector representing the log-likelihood-trace of
the NormR model fit.}

\item{\code{eps}}{A \code{numeric{}} representing preset threshold for NormR fit
convergence.}

\item{\code{lnposteriors}}{A \code{k}-dimensional \code{matrix()} containing ln
posterior probabilities for every unique \code{n} 2-tupel in counts given
each \code{theta} and \code{mixtures}. Use accessor
\code{getPosteriors()} to get the complete posterior matrix.}

\item{\code{lnenrichment}}{A \code{numeric()}-vector of normalized ln enrichment over
component \code{B} for each unique count-tupel. Use accessor
\code{\link{getEnrichment}} to retrieve enrichment for original count sequence.}

\item{\code{lnpvals}}{A \code{numeric()}-vector containing ln P-values every unique
2-tupel in counts given \code{theta} of \code{B}.}

\item{\code{filteredT}}{A \code{integer()}-vector giving indices of P-values to be
considered for FDR correction. These indices have been obtainted by filtering
P-values using the T method.}

\item{\code{lnqvals}}{A \code{numeric()}-vector containing ln q-values. These are
P-values corrected for multiple testing using Storey's method.}

\item{\code{classes}}{A \code{integer()}-vector giving class assignments based on
model fit. For \code{obj@type == "enrichR"}, this vector contains either
\code{NA} (not enriched) or \code{1} (enriched). For \code{obj@type ==
"diffR"}, this vector contains \code{NA} (unchanged), \code{1} (differential
in ChIP-seq 1) and \code{2} (differential in ChIP-seq 2). For
\code{obj@type == "regimeR"}, this vector contains \code{NA} (not enriched)
and an arbitary number of enrichment class \code{>= 1}.}
}}
\seealso{
\code{\link{normr-methods}} for the functions that produce
this object
}

