% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wrappers.R
\name{processByChromosome}
\alias{processByChromosome}
\title{Process a list of bam files with a number of processes.}
\usage{
processByChromosome(bam.files, gr, procs, bamsignals.function, mapqual,
  shift = 0, paired.end = "ignore", verbose = F)
}
\arguments{
\item{bam.files}{A list of strings specifying file system locations for bam
files to count.}

\item{gr}{A\link{GenomicRanges}-object giving the genomic intervals to count
in.}

\item{procs}{Number of threads to use in parallel::mclapply()}

\item{bamsignals.function}{The corresponding counting function in bamsignals
to call.}

\item{mapqual}{discard reads with mapping quality strictly lower than this
parameter. The value 0 ensures that no read will be discarded, the value 254
that only reads with the highest possible mapping quality will be considered.}

\item{shift}{shift the read position by a user-defined number of basepairs.
This can be handy in the analysis of chip-seq data.}

\item{paired.end}{a character string indicating how to handle paired-end
reads. If \code{paired.end!="ignore"} then only first reads in proper mapped
pairs will be consider (i.e. in the flag of the read, the bits in the mask
66 must be all ones). If \code{paired.end=="midpoint"} then the midpoint of a
fragment is considered, where \code{mid = fragment_start + int(abs(tlen)/2)},
and where tlen is the template length stored in the bam file. For even tlen,
the given midpoint will be moved of 0.5 basepairs in the 3' direction.
If \code{paired.end=="extend"} then the whole fragment is treated
as a single read.}

\item{verbose}{A logical value indicating whether verbose output is desired}
}
\value{
a list of length(bam.files) with bamsignals.function results
}
\description{
Process a list of bam files with a number of processes.
}

