% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormRFit.R
\docType{methods}
\name{writeEnrichment,NormRFit-method}
\alias{writeEnrichment}
\alias{writeEnrichment,NormRFit-method}
\title{describeIn NormRFit Writes the calculated enrichment of a NormRFit to disk.}
\usage{
\S4method{writeEnrichment}{NormRFit}(obj, gr = NULL, filename = NULL)
}
\arguments{
\item{obj}{A NormRFit object}
}
\description{
describeIn NormRFit Writes the calculated enrichment of a NormRFit to disk.
}

